DEV: Move residual error class to the proper folder
