+++ src/mdestimates/__init__.py
@@ -3,5 +3,5 @@ from .error_estimates import ErrorEstimate
 from mdestimates.pressure_reconstruction.pressure_reconstruction import PressureReconstruction
 from mdestimates.velocity_reconstruction.velocity_reconstruction import VelocityReconstruction
 from mdestimates.error_evaluation.diffusive_error import DiffusiveError
-from mdestimates.pressure_reconstruction.residual_error import ResidualError
+from mdestimates.error_evaluation.residual_error import ResidualError
 from .grid_rotation import RotatedGrid
